Implement new Wormhole telemetry
