+++ device/api/umd/device/arc_telemetry_reader.h
@@ -5,7 +5,10 @@
  */
 #pragma once
 
+#include <map>
+
 #include "umd/device/tt_device/tt_device.h"
+#include "umd/device/tt_xy_pair.h"
 
 namespace tt::umd {
 
@@ -13,14 +16,41 @@ class ArcTelemetryReader {
 public:
     virtual ~ArcTelemetryReader() = default;
 
-    virtual uint32_t read_entry(const uint8_t telemetry_tag) = 0;
+    uint32_t read_entry(const uint8_t telemetry_tag);
 
-    virtual bool is_entry_available(const uint8_t telemetry_tag) = 0;
+    bool is_entry_available(const uint8_t telemetry_tag);
 
     static std::unique_ptr<ArcTelemetryReader> create_arc_telemetry_reader(TTDevice* tt_device);
 
 protected:
     ArcTelemetryReader(TTDevice* tt_device);
+
+    virtual void get_telemetry_address() = 0;
+
+    void initialize_telemetry();
+
+    // Address of the telemetry table struct on ARC core.
+    uint32_t telemetry_table_addr;
+
+    // Number of entries in the telemetry table.
+    uint32_t entry_count;
+
+    // After entry_count the telemetry contains entry_count structs of TelemetryTagEntry.
+    // Each struct contains tag and offset. Tag represents what is represented by the value.
+    // Offset is the index of the telemetry value in the telemetry_values array.
+    struct TelemetryTagEntry {
+        uint16_t tag;
+        uint16_t offset;
+    };
+
+    // Address of the telemetry data on ARC core.
+    uint32_t telemetry_values_addr;
+
+    std::map<uint32_t, uint32_t> telemetry_values;
+    std::map<uint32_t, uint32_t> telemetry_offset;
+
+    // During initialization of telemetry, if the NOC0 is hung then we need to read the telemetry values from NOC1.
+    tt_xy_pair arc_core;
 
     TTDevice* tt_device;
 };
