+++ device/api/umd/device/wormhole_implementation.h
@@ -268,6 +268,8 @@ constexpr std::array<std::pair<CoreType, uint64_t>, 5> NOC1_CONTROL_REG_ADDR_BAS
      {CoreType::ARC, 0xFFFB30000}}};
 static const uint64_t NOC_NODE_ID_OFFSET = 0x2C;
 
+static const uint64_t ARC_RESET_UNIT_BASE_ADDR = 0x880030000;
+
 static const size_t tensix_translated_coordinate_start_x = 18;
 static const size_t tensix_translated_coordinate_start_y = 18;
