+++ tests/blackhole/test_arc_messages_bh.cpp
@@ -7,6 +7,7 @@
 #include "gtest/gtest.h"
 #include "umd/device/arc_messenger.h"
 #include "umd/device/blackhole_arc_telemetry_reader.h"
+#include "umd/device/blackhole_implementation.h"
 #include "umd/device/cluster.h"
 #include "umd/device/types/blackhole_arc.h"
